refactor: move about page into feature slice
