+++ src/app/(marketing)/about/page.tsx
@@ -1,4 +1,4 @@
-import MainSection from "./components/MainSection";
+import MainSection from "@/features/about/components/MainSection";
 
 export default function AboutPage() {
   return <MainSection />;
