+++ src/features/about/components/AchievementsSection.tsx
@@ -1,6 +1,6 @@
-import achievements from "@/data/achievements";
+import achievements from "@/features/about/data/achievements";
 
-import type { Achievement } from "@/data/achievements";
+import type { Achievement } from "@/features/about/data/achievements";
 
 export default function AchievementsSection() {
   return (
