+++ src/features/about/components/ChainOfThought.tsx
