+++ src/features/about/components/ExperienceSection.tsx
@@ -1,6 +1,6 @@
-import experiences from "@/data/experience";
+import experiences from "@/features/about/data/experience";
 
-import type { Experience } from "@/data/experience";
+import type { Experience } from "@/features/about/data/experience";
 
 export default function ExperienceSection() {
   return (
