+++ src/features/about/components/MainSection.tsx
