+++ src/features/about/components/QuoteSection.tsx
